Refactor Hero component to utilize HeroButtons for improved localization handling and code organization
