+++ src/components/sections/hero/hero-buttons.module.css
@@ -1,0 +1,9 @@
+.button:hover .arrow {
+  transform: translateY(3px);
+}
+
+.arrow {
+  transition-property: transform;
+  transition-duration: 500ms;
+  transition-timing-function: ease-in-out;
+}
